Move SelfAttention and Block classes to attention.py
